% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{clean_noise}
\alias{clean_noise}
\title{Noise cleaning wrapper}
\usage{
clean_noise(dataset, method, class_attr = "Class", ...)
}
\arguments{
\item{dataset}{we want to clean noisy instances on}

\item{method}{selected method of noise cleaning}

\item{class_attr}{\code{character}. Indicates the class attribute or
attributes from \code{dataset}. Must exist in it.}

\item{...}{Further arguments for \code{method}}
}
\value{
The treated dataset (either with noisy instances replaced or erased)
}
\description{
Noise cleaning wrapper
}
\examples{
library("amendr")

clean_noise(iris, method = "AENN", class_attr = "Species", k = 3)
clean_noise(iris, method = "AENN", class_attr = "Species")

}
