% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{discretize}
\alias{discretize}
\title{Discretization wrapper}
\usage{
discretize(dataset, method, class_attr = "Class", ...)
}
\arguments{
\item{dataset}{we want to perform discretization on}

\item{method}{selected method of discretization}

\item{class_attr}{\code{character}. Indicates the class attribute or
attributes from \code{dataset}. Must exist in it.}

\item{...}{Further arguments for \code{method}}
}
\value{
The discretized dataset
}
\description{
Discretization wrapper
}
\examples{
library("amendr")
library("imbalance")
library("magrittr")
data(iris0)

super_iris <- iris0 \%>\% discretize(method = "chi2", class_attr = "Class")
super_iris <- iris0 \%>\% discretize(method = "ameva", class_attr = "Class"
super_iris <- iris0 \%>\% discretize(method = "equalwidth", class_attr = "Class")

}
