% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValues.R
\name{impute_missing}
\alias{impute_missing}
\title{Missing values imputation wrapper}
\usage{
impute_missing(dataset, method, ...)
}
\arguments{
\item{dataset}{we want to impute missing values on}

\item{method}{selected method of missing values imputation}

\item{...}{Further arguments for \code{method}}
}
\value{
The treated dataset (either with noisy instances replaced or erased)
}
\description{
Missing values imputation wrapper
}
\examples{
library("amendr")
data(africa, package = "Amelia")
data(nhanes, package = "mice")

super_nhanes <- impute_missing(nhanes, "gibbs_sampling")
# Use a different method for every column
impute_methods <- c("pmm", "midastouch", "norm_nob", "norm_boot")
super_nhanes <- impute_missing(nhanes, "gibbs_sampling", imputation = impute_methods)
super_nhanes <- impute_missing(nhanes, "central_imputation")
super_africa <- impute_missing(africa, "knn_imputation")
# Execute knn imputation with non default value for k
super_africa <- impute_missing(africa, "knn_imputation", k = 5)
super_africa <- impute_missing(africa, "expect_maximization", exclude = "country")
super_africa <- impute_missing(africa, "rf_imputation")
super_africa <- impute_missing(africa, "rf_imputation", num_iterations = 15,
                               num_trees = 200, bootstrap = FALSE)

}
