% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{Normalization wrapper}
\usage{
normalization(dataset, method = normalizationMethods, class_attr = "Class",
  ...)
}
\arguments{
\item{dataset}{we want to perform normalization on}

\item{method}{selected method of normalization}

\item{classAttr}{\code{character}. Indicates the class attribute from
\code{dataset}. Must exist in it.}
}
\value{
The normalized dataset
}
\description{
Normalization wrapper
}
\examples{
library("amendr")
data(iris0)

super_iris <- iris0 \%>\% normalization(method = "min-max", class_attr = "Class")
super_iris <- iris0 \%>\% normalization(method = "min-max", normalization = "row")

}
