% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amendr.R
\name{options}
\alias{options}
\title{Prints options for a certain preprocessing method}
\usage{
options(preprocess, method = NULL)
}
\arguments{
\item{preprocess}{Possible preprocessing: \code{'clean_noise'}}

\item{method}{For the preprocessing method}
}
\value{
Prints options for the selected preprocessing
}
\description{
Prints options for a certain preprocessing method
}
