% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers_clean}
\alias{outliers_clean}
\title{Outliers cleaning wrapper}
\usage{
outliers_clean(dataset, method, ...)
}
\arguments{
\item{dataset}{we want to clean outliers on}

\item{method}{selected method to clean outliers. Possbilities are:
\itemize{
 \item{"univariate"}{detects outliers column by column (an outlier will be an
  abnormal value inside a column) and fills them with mean or median of the
  corresponding column}
 \item{"multivariate"}{detects outliers using a multicolumn approach, so that
  an outlier will be a whole observation (row). And deletes those
  observations}
}}

\item{...}{further arguments for the method}
}
\value{
The dataset without outliers
}
\description{
Outliers cleaning wrapper
}
\examples{
library("amendr")
data(ecoli1, package = "imbalance")
data(iris0, package = "imbalance")

super_ecoli <- outliers_clean(ecoli1, method = "multivariate", type = "adj")
super_ecoli <- outliers_clean(ecoli1, method = "multivariate", type = "quan")

# Use mean as method to substitute outliers
outliers_clean(iris0, method = "univariate", type = "z", prob = 0.9, fill = "mean")
# Use median as method to substitute outliers
outliers_clean(iris0, method = "univariate", type = "z", prob = 0.9, fill = "median")
# Use chi-sq instead of z p-values
outliers_clean(iris0, method = "univariate", type = "chisq", prob = 0.9, fill = "median")
# Use interquartilic range instead (lim argument is mandatory when using it)
outliers_clean(iris0, method = "univariate", type = "iqr", lim = 0.9, fill = "median")

}
