% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartdata.R
\docType{package}
\name{smartdata}
\alias{smartdata}
\title{smartdata: A package to ease data preprocessing tasks}
\description{
Provides a pipe interface that integrates a collection of the most used data
preprocessing libraries, providing oversampling, instance and feature
selection, normalization, discretization, space transformation and
outliers/noise/missing values treatment
}
\section{Method to do oversampling \code{\link{oversample}}}{
NA
}

\section{Method to do instance selection \code{\link{instance_selection}}}{
NA
}

\section{Method to do feature selection \code{\link{feature_selection}}}{
NA
}

\section{Method to do normalization \code{\link{normalize}}}{
NA
}

\section{Method to do discretization \code{\link{discretize}}}{
NA
}

\section{Method to do space transformation \code{\link{space_transformation}}}{
NA
}

\section{Method to treat outliers \code{\link{clean_outliers}}}{
NA
}

\section{Method to treat missing values \code{\link{impute_missing}}}{
NA
}

\section{Method to treat noise \code{\link{clean_noise}}}{
NA
}

