% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartdata.R
\name{which_options}
\alias{which_options}
\title{Prints options for a certain preprocessing method}
\usage{
which_options(preprocess, method = NULL)
}
\arguments{
\item{preprocess}{Possible preprocessing: \code{'clean_noise'}}

\item{method}{For the preprocessing method}
}
\value{
Prints options for the selected preprocessing
}
\description{
Prints options for a certain preprocessing method
}
\examples{

which_options("clean_noise", method = "edgeWeight")
which_options("clean_noise", method = "ENG")
which_options("impute_missing", method = "gibbs_sampling")

}
