% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amendr.R
\docType{package}
\name{amendr}
\alias{amendr}
\alias{amendr-package}
\title{amendr: A package to ease data preprocessing tasks}
\description{
Provides a pipe interface that integrates a collection of the most used data
preprocessing libraries, providing oversampling, instance and feature
selection, normalization, discretization, space transformation and
outliers/noise/missing values treatment
}
\section{Method to do oversampling \code{\link{}}}{
NA
}

\section{Method to do instance selection \code{\link{}}}{
NA
}

\section{Method to do feature selection \code{\link{}}}{
NA
}

\section{Method to do normalization \code{\link{}}}{
NA
}

\section{Method to do discretization \code{\link{}}}{
NA
}

\section{Method to do space transformation \code{\link{}}}{
NA
}

\section{Method to treat outliers \code{\link{}}}{
NA
}

\section{Method to treat missing values \code{\link{}}}{
NA
}

\section{Method to treat noise \code{\link{}}}{
NA
}

