% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{discretize}
\alias{discretize}
\title{Discretization wrapper}
\usage{
discretize(dataset, method, class_attr = "Class", ...)
}
\arguments{
\item{dataset}{we want to perform discretization on}

\item{method}{selected method of discretization}

\item{class_attr}{\code{character}. Indicates the class attribute or
attributes from \code{dataset}. Must exist in it.}

\item{...}{Further arguments for \code{method}}
}
\value{
The discretized dataset
}
\description{
Discretization wrapper
}
\examples{
library("amendr")

discretize(iris, method = "chi-merge", class_attr = "Species")
discretize(iris, method = "chi-merge", class_attr = "Species", alpha = 0.7)
discretize(iris, method = "chi2", class_attr = "Species", alpha = 0.7, delta = 0.1)
discretize(iris, method = "chi2", class_attr = "Species")
discretize(iris, method = "extended-chi2", class_attr = "Species")
discretize(iris, method = "ameva", class_attr = "Species")
discretize(iris, method = "CAIM", class_attr = "Species")
discretize(iris, method = "CACC", class_attr = "Species")
discretize(iris, method = "equalwidth", class_attr = "Species", num_bins = nrow(iris) / 2)

}
