% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{discretize}
\alias{discretize}
\title{Discretization wrapper}
\usage{
discretize(dataset, method, exclude = NULL, ...)
}
\arguments{
\item{dataset}{we want to perform discretization on}

\item{method}{selected method of discretization}

\item{exclude}{\code{character}. Vector of attributes to exclude from the
discretization}

\item{...}{Further arguments for \code{method}}
}
\value{
The discretized dataset
}
\description{
Discretization wrapper
}
\examples{
library("amendr")

super_iris <- discretize(iris, method = "chi_merge", exclude = c("Species", "Sepal.Length"))
super_iris <- discretize(iris, method = "chi_merge", exclude = "Species", alpha = 0.7)
super_iris <- discretize(iris, method = "chi2", "Species", alpha = 0.7, delta = 0.1)
super_iris <- discretize(iris, method = "chi2", exclude = "Species")
super_iris <- discretize(iris, method = "extended_chi2", exclude = "Species")
super_iris <- discretize(iris, method = "ameva", exclude = "Species")
super_iris <- discretize(iris, method = "CAIM", exclude = "Species")
super_iris <- discretize(iris, method = "CACC", exclude = "Species")
super_iris <- discretize(iris, method = "equalwidth", "Species", num_bins = nrow(iris) / 2)
super_iris <- discretize(iris, method = "equalfreq", "Species", num_bins = nrow(iris) / 2)

}
