% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{feature_selection}
\alias{feature_selection}
\title{Feature selection wrapper}
\usage{
feature_selection(dataset, method, class_attr = "Class", ...)
}
\arguments{
\item{dataset}{we want to do feature selection on}

\item{method}{selected method of feature selection}

\item{class_attr}{\code{character}. Indicates the class attribute or
attributes from \code{dataset}. Must exist in it.}

\item{...}{Further arguments for \code{method}}
}
\value{
The treated dataset (either with noisy instances replaced or erased)
}
\description{
Feature selection wrapper
}
\examples{
library("amendr")
data(ecoli1, package = "imbalance")

super_iris <- feature_selection(iris, "Boruta", class_attr = "Species")
super_iris <- feature_selection(iris, "chi_squared",
                                class_attr = "Species", num_attrs = 3)
# Pick 3 attributes from the continuous ones
super_iris <- feature_selection(ecoli1, "information_gain",
                                class_attr = "Class", num_attrs = 3)
super_iris <- feature_selection(ecoli1, "gain_ratio",
                                class_attr = "Class", num_attrs = 3)
super_iris <- feature_selection(ecoli1, "sym_uncertainty",
                                class_attr = "Class", num_attrs = 3)

}
