% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_selection.R
\name{instance_selection}
\alias{instance_selection}
\title{Instance selection wrapper}
\usage{
instance_selection(dataset, method = instSelectionMethods,
  classAttr = "Class")
}
\arguments{
\item{dataset}{we want to perform an instance selection on}

\item{method}{selected method of instance selection}

\item{classAttr}{\code{character}. Indicates the class attribute from
\code{dataset}. Must exist in it.}
}
\value{
A filtered dataset with same shape as the input to the function
}
\description{
Instance selection wrapper
}
\examples{
library("amendr")
data(iris0)

super_iris <- iris0 \%>\% instance_selection(method = "CNN")
}
