% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instanceSelection.R
\name{instance_selection}
\alias{instance_selection}
\title{Instance selection wrapper}
\usage{
instance_selection(dataset, method, class_attr = "Class", ...)
}
\arguments{
\item{dataset}{we want to perform an instance selection on}

\item{method}{selected method of instance selection}

\item{class_attr}{\code{character}. Indicates the class attribute from
\code{dataset}. Must exist in it}

\item{...}{Further arguments for \code{method}}
}
\value{
A filtered dataset with same shape as the input to the function
}
\description{
Instance selection wrapper
}
\examples{
library("amendr")
library("imbalance")
library("magrittr")
data(iris0)

super_iris <- iris0 \%>\% instance_selection(method = "CNN")
}
