% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbalance.R
\name{oversample}
\alias{oversample}
\title{Oversampling wrapper}
\usage{
oversample(dataset, method, ratio = 0.8, class_attr = "Class", ...)
}
\arguments{
\item{dataset}{we want to perform oversampling on}

\item{method}{selected method of oversampling}

\item{class_attr}{\code{character}. Indicates the class attribute or
attributes from \code{dataset}. Must exist in it.}

\item{...}{Further arguments for \code{method}}
}
\value{
An oversampled dataset
}
\description{
Oversampling wrapper
}
\examples{
library("amendr")
library("imbalance")
library("magrittr")
data(iris0)

super_iris <- iris0 \%>\% oversample(method = "MWMOTE", class_attr = "Class")
super_iris <- iris0 \%>\% discretize(method = "SMOTE", class_attr = "Class"
super_iris <- iris0 \%>\% discretize(method = "PDFOS", class_attr = "Class")

}
