% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceTransformation.R
\name{space_transformation}
\alias{space_transformation}
\title{Space transformation wrapper}
\usage{
space_transformation(dataset, method, class_attr = "Class", ...)
}
\arguments{
\item{dataset}{we want to do space transformation on}

\item{method}{selected method of space transformation}

\item{class_attr}{\code{character}. Indicates the class attribute or
attributes from \code{dataset}. Must exist in it.}

\item{...}{Further arguments for \code{method}}
}
\value{
The transformed dataset
}
\description{
Space transformation wrapper
}
\examples{
library("amendr")
data(ecoli1, package = "imbalance")


super_iris <- space_transformation(ecoli1, "lle_knn", k = 3, num_features = 3)
\dontrun{
super_iris <- space_transformation(ecoli1, "lle_epsilon", epsilon = 0.99, num_features = 3)
}

}
