% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceTransformation.R
\name{space_transformation}
\alias{space_transformation}
\title{Space transformation wrapper}
\usage{
space_transformation(dataset, method, exclude = NULL, ...)
}
\arguments{
\item{dataset}{we want to do space transformation on}

\item{method}{selected method of space transformation}

\item{exclude}{\code{character}. Vector of attributes to exclude from the
space transformation process}

\item{...}{Further arguments for \code{method}}
}
\value{
The transformed dataset
}
\description{
Space transformation wrapper
}
\examples{
library("amendr")
data(ecoli1, package = "imbalance")
data(AntibioticSmall, package = "adaptiveGPCA")
antibiotics <- data.frame(AntibioticSmall$X)

super_iris <- space_transformation(ecoli1, "lle_knn", k = 3, num_features = 2,
                                   regularization = 1, exclude = c("Mcg", "Alm1"))
\dontrun{
super_iris <- space_transformation(ecoli1, "lle_epsilon", epsilon = 0.99, num_features = 3)
}
super_iris <- space_transformation(antibiotics, "adaptative_gpca", similarity = AntibioticSmall$Q,
                                   num_features = 2)

}
